pyproject.toml
[project]
name = "kvikio-zarr-v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["kvikio-cu12>=25.2.0", "nvtx>=0.2.11", "zarr", "ipython", "xarray"]

[tool.uv.sources]
# zarr = { git = "https://github.com/akshaysubr/zarr-python", rev = "gpu-codecs" }  # need to push tags here
zarr = { git = "https://github.com/TomAugspurger/zarr-python", rev = "gpu-codecs" }  # need to push tags here
# xarray = { git = "https://github.com/pydata/xarray", rev = "fix-cupy"}
xarray = { git = "https://github.com/TomAugspurger/xarray", rev = "force-cpu-fix-cupy" }

[dependency-groups]
dev = [
    "ruff>=0.9.7",
]

[tool.uv]
override-dependencies = [
    # "zarr @ git+https://github.com/akshaysubr/zarr-python@gpu-codecs",
    "zarr==3.0.4",
    "nvidia-nvcomp-cu12==4.2.0.14",
    # "xarray @ ../xarray",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B", "UP", "N", "S", "ERA", "PD", "EXE", "PGH"]
